# lib/api-client-react/src/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Village Swimming Pool Reservation System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterInput {
  firstName: string;
  lastName: string;
  houseNumber: string;
  phone: string;
  email: string;
  username: string;
  password: string;
  role?: string;
}

export interface LoginInput {
  usernameOrEmail: string;
  password: string;
  rememberMe?: boolean;
}

export interface User {
  id: number;
  firstName: string;
  lastName: string;
  houseNumber: string;
  phone: string;
  email: string;
  username: string;
  role: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface UserUpdate {
  firstName?: string;
  lastName?: string;
  houseNumber?: string;
  phone?: string;
  email?: string;
  role?: string;
}

export interface UserListResponse {
  users: User[];
  total: number;
  page: number;
  totalPages: number;
}

export interface ResetPasswordInput {
  newPassword: string;
}

export interface ReservationInput {
  date: string;
  startTime: string;
  endTime: string;
  numberOfPeople: number;
  /** @nullable */
  notes?: string | null;
}

export interface ReservationUpdate {
  status?: string;
  /** @nullable */
  notes?: string | null;
  numberOfPeople?: number;
}

export interface Reservation {
  id: number;
  userId: number;
  date: string;
  startTime: string;
  endTime: string;
  numberOfPeople: number;
  status: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  user: User;
}

export interface ReservationListResponse {
  reservations: Reservation[];
  total: number;
  page: number;
  totalPages: number;
}

export interface TimeSlot {
  startTime: string;
  endTime: string;
  available: boolean;
  currentPeople: number;
  maxPeople: number;
}

export interface Settings {
  id: number;
  bookingEnabled: boolean;
  openTime: string;
  closeTime: string;
  maxPeoplePerSlot: number;
  maxAdvanceDays: number;
  slotDurationMinutes: number;
  maintenanceMode: boolean;
  /** @nullable */
  maintenanceMessage?: string | null;
}

export interface SettingsUpdate {
  bookingEnabled?: boolean;
  openTime?: string;
  closeTime?: string;
  maxPeoplePerSlot?: number;
  maxAdvanceDays?: number;
  slotDurationMinutes?: number;
  maintenanceMode?: boolean;
  /** @nullable */
  maintenanceMessage?: string | null;
}

export interface AdminStats {
  totalMembers: number;
  totalReservations: number;
  todayReservations: number;
  monthReservations: number;
  upcomingReservations: number;
  cancelledThisMonth: number;
}

export interface MemberStats {
  totalReservations: number;
  upcomingCount: number;
  cancelledCount: number;
  thisMonthCount: number;
}

export interface MonthlyStatItem {
  month: string;
  count: number;
  cancelled: number;
}

export interface TopUser {
  id: number;
  firstName: string;
  lastName: string;
  houseNumber: string;
  reservationCount: number;
}

export type ListUsersParams = {
search?: string;
page?: number;
limit?: number;
};

export type ListReservationsParams = {
date?: string;
startDate?: string;
endDate?: string;
status?: string;
userId?: number;
page?: number;
limit?: number;
};

export type GetAvailableSlotsParams = {
date: string;
};

export type GetMyReservationsParams = {
status?: string;
page?: number;
limit?: number;
};

export type GetMonthlyStatsParams = {
year?: number;
month?: number;
};